logc: add log_would_log function
The idea is to allow user of LogC to check if preparation of output even
makes sense.

This should be used when user wants to run some time consuming data
collection and assembly to put together message. Such process can be
just skipped if it would not be used anyway.