logc: adding existing output replaces it instead of append
This change is done to expand possibilities of output control. It makes
no sense to have multiple outputs to same file (and could also in some
cases lead to problems). Thanks to that we can use object address as
representation of given object. Adding existing object won't append it
but instead modifies existing one.