logc: fix how level condition work and add origin setting
There was a collision in design of conditions that we can either check
for level of output or for level of message. It makes much more sense to
have it depending on level of message so it was converted to that. That
also fixes usage with colors.

That opened up question on how to handle other case when we do want to
check for setting of logger rather message. In the end it seems like the
only reason why we would want to do this is to print additional trace
info and that is origin of message. Because of that this replaces that
usage with dedicated option that does exactly that.

In future we might return to original design and introduce both
condition. Even in such case it makes sense to not tie origin print to
level of logger as with not yet implemented design of subordinate
loggers it is going to be less clear if user really wants to trace or
just wants higher level to see also subordinate messages.