Remove level names and instead define them as contant
This is more generic approach as string representation of level is not
in any way hard-coded this way.
Second and important change is in a way how empty conditions work.
Difference is that empty condition now considers sub-conditions as
non-empty if they expand to something and as empty if they don't. In
effect this implementation won't dive in to sub-conditions.