Change how message level works and is handled
This makes message and log level handling well defined. The rule is that
message level/type has some offset assigned relative to zero and log
level just shifts the limit for these to be used. This was required for
output level setting as well as for chaining and syslog in future.

How it works now is that all offsets are added together, that is at the
moment just level from log and level from output. This sum is then
compared with message level. Only messages with higher or equal level
are printed.

The message level can be used to format message while log and output
levels can't be and are used only to filter messages.