logc: various fixes and cleanups
This improves and simplifies some of the tests as well as moves level
handling functions to separate file.

Another notable change is also that _log no longer uses log_would_log to
check if it should run. Instead it just runs. The reason is that most of
the preparations from beginning of the function were removed here.