logc: implement syslog support
This adds support for output to syslog. No advanced configuration except
of format change  is supported as standard functions such as setlogmask
or openlog can be used instead.
Code considers syslog as single output target so only one format can be
specified for syslog output and message is sent to syslog only once.