configure: make gperf not essential for distribution
Distribution archives do not essentially need gperf. This implements
that. It is easy to ensure that distribution archive contains generated
files instead of source. The only hack here is disable of check for
gperf in distribution archive. The problem is how to detect distribution
archive in configure. In the end bootstrap script is not included so it
is used to detect source version versus distribution archive.