logc_argp: add --syslog and introduce dedicated  parser for daemons
This adds --syslog option to enable logging to syslog. It disables
logging to stderr. Effectively this is redirect of logs to syslog.

In case of daemons logging to syslog is explicitly disabled so we want
to rather disable sometimes logging to stderr. Because of that the new
dedicated argument parser was introduced with option --no-stderr.