logc: add output locking
There is problem with output from multiple threads and subprocesses
being mixed together. This should prevent this.

This is not ideal as we are using file objects and usage of file
descriptor locking together with file objects is discouraged by manual.
This is probably going required move away from file object directly to
file descriptors.