logc: try to optimize debug messages
This introduces pretty reasonable assumption that trace level messages
are called almost every time they are encountered in the code (they are
in likely branches) while error messages are not (they are in unlikely
branches).
In the LogC code it is more efficient to check if log would generate any
output over just executing _log function without actually generating any
output. The reason for this is because we prepare and check outputs as
well as that we prepare full user message before we go trough outputs.

In the end it should be much better in most cases to check if level
would generate any output for messages in more likely branches and at
the same time to just go trough logging for messages in unlikely
branches.