Convert macros to use linux coding style for argument names
The Linux coding style we are vaguely following uses lower case for
argument names for macros. We should do that as well. It can be used to
recognize what is global symbol and what is just an argument.

This also converts utility macro 'std_fatal' to statements that is
providing resulting value. This way it can be used also when we want to
know result of function. In such case they serve only to protect against
unlikely error.