logc_asserts: add more informative asserts
This adds new header with in the end only few macros that are expected
to be used by user and those are:
* assert_eq
* assert_ne
* assert_lt
* assert_le
* assert_gt
* assert_ge

These new macros try to identify basic types and compare them in more
informed way compared to plain assert. The advantage is that we have
access to products of left and right sides evaluations and thus we can
report them.

This is handy when you want to assert on very unlikely case but need
values just to know why assert actually happened.