build: tweak bootstrap detection for gperf remove gen-gperf target
The bootstrap detection has to be relative to source directory but
configure is run in build directory where bootstrap file is not present.

The removal of the gen-gperf target is reasoned as not being that
useful. We can generate separate C files from gperf just by targeting
output file and there is no general use for generating all C files from
gperf.