Add the initial template based on existing projects
This is heavily improved common code that we use in multiple projects.
To have minimal project prepared for fork should give us way faster C
projects creation. Another advantage is that we can merge new changes to
existing projects and that way ensure that fixes to common template are
reliably propagated.