logc_argp: replace no-stderr with no-syslog for daemon logs
The ability to disable stderr logging is not exactly needed most of the
time. It is possible to close or redirect stderr to /dev/null to just
discard stderr logs and that is common on process daemonization.
On the other hand the ability to disable syslog that is in default
enabled for daemon logs is required.