Add support to include LogC as Meson's subproject
This way it is easy to fulfill the dependencies of project using
Meson's subproject.

The project has to specify wrap file 'subprojects/logc.wrap' with
following content:

  [wrap-git]
  url = https://gitlab.nic.cz/turris/logc.git
  revision = master

Then dependency on LogC automatically fallbacks to the subproject. The
support for other libraries needs a little tweak but is easy as well:

  logc = dependency('logc')
  logc_argp = dependency('logc_argp', fallback: ['logc', 'logc_argp_dep'])
  logc_config = dependency('logc_config', fallback: ['logc', 'logc_config_dep'])