meson: do not enable tests in case it is subproject
The inclusion of subprojects is mostly to get easy access to the library
without having to install it. It makes in most cases no sense to run
tests of subproject when we are not developing it.

In general I am sure that including tests from subproject makes a lot of
sense, for example when we split tests and application to the separate
repositories. At the same time we are really not doing that here.